wait for mount point device number to change
